requirements.txt
aiofiles==24.1.0
altair==5.4.1
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
decorator==5.1.1
etelemetry==0.3.1
executing==2.1.0
faiss-cpu==1.8.0.post1
filelock==3.16.1
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
ipython==8.27.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lightrag==0.1.0b6
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.8.2
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
ollama==0.3.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.28.2
prov==2.0.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.39.0
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tools==0.1.9
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
